requirements.txt
Django~=3.2.7
djangorestframework~=3.12.4
django-rest-auth~=0.9.5
gunicorn~=20.1.0